chore: hub, movementInfo 서버 주석 추가
