+++ com.tech-nova.hub/src/main/java/com/tech_nova/hub/application/service/HubService.java
@@ -3,96 +3,54 @@ package com.tech_nova.hub.application.service;
 import com.tech_nova.hub.application.dtos.req.HubRequestDto;
 import com.tech_nova.hub.application.dtos.res.HubClientResponseDto;
 import com.tech_nova.hub.application.dtos.res.HubResponseDto;
-import com.tech_nova.hub.domain.model.Hub;
-import com.tech_nova.hub.domain.repository.HubRepository;
-import com.tech_nova.hub.presentation.exception.HubNotFoundException;
-import com.tech_nova.hub.presentation.exception.MasterRoleRequiredException;
-import com.tech_nova.hub.presentation.exception.RoleNotAllowedException;
 import java.util.List;
 import java.util.UUID;
-import lombok.RequiredArgsConstructor;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.CachePut;
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.cache.annotation.Caching;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@RequiredArgsConstructor
-public class HubService {
+public interface HubService {
 
-  private final HubRepository hubRepository;
+  /**
+   * 허브 생성
+   *
+   * @param hubRequestDto 생성할 허브의 세부 사항을 담고 있는 DTO
+   * @param userId        요청을 하는 사용자 ID
+   * @param role          요청을 하는 사용자의 역할
+   * @return 생성된 허브의 응답 DTO
+   */
+  HubResponseDto createHub(HubRequestDto hubRequestDto, UUID userId, String role);
 
-  @CachePut(cacheNames = "HubCache", key = "#result.hubId")
-  @Transactional
-  public HubResponseDto createHub(HubRequestDto hubRequestDto, UUID userId, String role) {
+  /**
+   * 허드 단일 조회
+   *
+   * @param hubId 조회할 허브의 ID
+   * @param role  요청을 하는 사용자의 권한
+   * @return 조회된 허브의 응답 DTO
+   */
+  HubResponseDto getHub(UUID hubId, String role);
 
-    validateMasterRole(role);
+  /**
+   * 허브 업데이트
+   *
+   * @param hubId         업데이트할 허브의 ID
+   * @param hubRequestDto 업데이트된 허브의 세부 사항을 담고 있는 DTO
+   * @param userId        요청을 하는 사용자 ID
+   * @param role          요청을 하는 사용자의 권한
+   * @return 업데이트된 허브의 응답 DTO
+   */
+  HubResponseDto updateHub(UUID hubId, HubRequestDto hubRequestDto, UUID userId, String role);
 
-    Hub hub = Hub.createHub(hubRequestDto, userId);
-    hubRepository.save(hub);
+  /**
+   * 허브 삭제
+   *
+   * @param hubId  삭제할 허브의 ID
+   * @param userId 요청을 하는 사용자 ID
+   * @param role   요청을 하는 사용자의 권한
+   */
+  void deleteHub(UUID hubId, UUID userId, String role);
 
-    return HubResponseDto.of(hub);
-  }
-
-  @Transactional(readOnly = true)
-  public HubResponseDto getHub(UUID hubId, String role) {
-
-    Hub hub = switch (role) {
-      case "MASTER" -> findHubById(hubId);
-      case "COMPANY", "DELIVERY", "HUB" -> findByHubIdAndDeletedFalse(hubId);
-      default -> throw new RoleNotAllowedException("유효하지 않은 권한입니다.");
-    };
-
-    return HubResponseDto.of(hub);
-  }
-
-  @Caching(
-      put = @CachePut(cacheNames = "HubCache", key = "#hubId"),
-      evict = @CacheEvict(cacheNames = "HubAllCache", allEntries = true)
-  )
-  @Transactional
-  public HubResponseDto updateHub(UUID hubId, HubRequestDto hubRequestDto, UUID userId, String role) {
-
-    validateMasterRole(role);
-
-    Hub hub = findHubById(hubId);
-    hub.updateHub(hubRequestDto, userId);
-
-    return HubResponseDto.of(hub);
-  }
-
-  @CacheEvict(cacheNames = {"HubCache", "HubAllCache"}, allEntries = true)
-  @Transactional
-  public void deleteHub(UUID hubId, UUID userId, String role) {
-
-    validateMasterRole(role);
-    findHubById(hubId).deleteHub(userId);
-  }
-
-  @Cacheable(cacheNames = "HubAllCache", key = "methodName")
-  @Transactional(readOnly = true)
-  public List<HubClientResponseDto> getHubList() {
-    return hubRepository.findAll()
-        .stream()
-        .map(HubClientResponseDto::of)
-        .toList();
-  }
-
-  private void validateMasterRole(String role) {
-    if (!"MASTER".equals(role)) {
-      throw new MasterRoleRequiredException("마스터 권한을 가진 사용자가 아닙니다.");
-    }
-  }
-
-  private Hub findHubById(UUID hubId) {
-    return hubRepository.findById(hubId)
-        .orElseThrow(() -> new HubNotFoundException("해당 허브가 존재하지 않습니다."));
-  }
-
-  private Hub findByHubIdAndDeletedFalse(UUID hubId) {
-    return hubRepository.findByHubIdAndIsDeletedFalse(hubId)
-        .orElseThrow(() -> new HubNotFoundException("해당 허브가 존재하지 않습니다."));
-  }
+  /**
+   * 모든 허브의 목록을 조회
+   *
+   * @return 허브 클라이언트 응답 DTO 목록
+   */
+  List<HubClientResponseDto> getHubList();
 }
